Rename `value_ir` to `ir`
