+++ src/frontend/mod.rs
@@ -38,7 +38,7 @@ impl Manager {
         // compiler::compile_modules(modules);
 
         let ir_modules =
-            compiler::value_ir::compile_modules(manager.0.modules.borrow().iter(), &entry)
+            compiler::ir::compile_modules(manager.0.modules.borrow().iter(), &entry)
                 .get_modules();
 
         compiler::target::compile_modules(&ir_modules);
